* Create View Component
