+++ BookStore.WebUI/Areas/Identity/Pages/Account/Logout.cshtml.cs
@@ -26,7 +26,6 @@ namespace BookStore.WebUI.Areas.Identity.Pages.Account
         public async Task<IActionResult> OnPost(string returnUrl = null)
         {
             await _signInManager.SignOutAsync();
-            HttpContext.Session.Clear();
             _logger.LogInformation("User logged out.");
             if (returnUrl != null)
             {
